+++ src/training/numba_desktop/simulate_learned_strategy.py
@@ -1,6 +1,6 @@
 import numpy as np
-from ..inverted_pendulum_simulator.src.inverted_pendulum import InvertedPendulum
-from ..inverted_pendulum_simulator.src.inverted_pendulum_visualizer import (
+from src.inverted_pendulum_simulator.src.inverted_pendulum import InvertedPendulum
+from src.inverted_pendulum_simulator.src.inverted_pendulum_visualizer import (
     InvertedPendulumVisualizer,
 )
 from .q_learning import QLearning
